+++ lib/ash_authentication/token_resource/actions.ex
@@ -70,20 +70,42 @@ defmodule AshAuthentication.TokenResource.Actions do
   def token_revoked?(resource, token, opts \\ []) do
     with :ok <- assert_resource_has_extension(resource, TokenResource),
          {:ok, domain} <- Info.token_domain(resource),
-         {:ok, is_revoked_action_name} <- Info.token_revocation_is_revoked_action_name(resource) do
-      resource
-      |> Query.new()
-      |> Query.set_context(%{
-        private: %{
-          ash_authentication?: true
-        }
-      })
-      |> Query.for_read(is_revoked_action_name, %{"token" => token}, opts)
-      |> Ash.read(domain: domain)
-      |> case do
-        {:ok, []} -> false
-        {:ok, _} -> true
-        _ -> false
+         {:ok, is_revoked_action_name} <- Info.token_revocation_is_revoked_action_name(resource),
+         action when not is_nil(action) <-
+           Ash.Resource.Info.action(resource, is_revoked_action_name) do
+      case action.type do
+        :action ->
+          resource
+          |> Ash.ActionInput.for_action(
+            is_revoked_action_name,
+            %{"token" => token},
+            Keyword.put(opts, :domain, domain)
+          )
+          |> Ash.run_action()
+          |> case do
+            {:ok, value} -> value
+            _ -> false
+          end
+
+        :read ->
+          resource
+          |> Query.new()
+          |> Query.set_context(%{
+            private: %{
+              ash_authentication?: true
+            }
+          })
+          |> Query.for_read(
+            is_revoked_action_name,
+            %{"token" => token},
+            Keyword.put(opts, :domain, domain)
+          )
+          |> Ash.read()
+          |> case do
+            {:ok, []} -> false
+            {:ok, _} -> true
+            _ -> false
+          end
       end
     end
   end
@@ -98,20 +120,42 @@ defmodule AshAuthentication.TokenResource.Actions do
   def jti_revoked?(resource, jti, opts \\ []) do
     with :ok <- assert_resource_has_extension(resource, TokenResource),
          {:ok, domain} <- Info.token_domain(resource),
-         {:ok, is_revoked_action_name} <- Info.token_revocation_is_revoked_action_name(resource) do
-      resource
-      |> Query.new()
-      |> Query.set_context(%{
-        private: %{
-          ash_authentication?: true
-        }
-      })
-      |> Query.for_read(is_revoked_action_name, %{"jti" => jti}, opts)
-      |> Ash.read(domain: domain)
-      |> case do
-        {:ok, []} -> false
-        {:ok, _} -> true
-        _ -> false
+         {:ok, is_revoked_action_name} <- Info.token_revocation_is_revoked_action_name(resource),
+         action when not is_nil(action) <-
+           Ash.Resource.Info.action(resource, is_revoked_action_name) do
+      case action.type do
+        :action ->
+          resource
+          |> Ash.ActionInput.for_action(
+            is_revoked_action_name,
+            %{"jti" => jti},
+            Keyword.put(opts, :domain, domain)
+          )
+          |> Ash.run_action()
+          |> case do
+            {:ok, value} -> value
+            _ -> false
+          end
+
+        :read ->
+          resource
+          |> Query.new()
+          |> Query.set_context(%{
+            private: %{
+              ash_authentication?: true
+            }
+          })
+          |> Query.for_read(
+            is_revoked_action_name,
+            %{"jti" => jti},
+            Keyword.put(opts, :domain, domain)
+          )
+          |> Ash.read()
+          |> case do
+            {:ok, []} -> false
+            {:ok, _} -> true
+            _ -> false
+          end
       end
     end
   end